pyproject.toml
[project]
name = "photoshare"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.0.0",
    "django>=5.2.1",
    "django-allauth>=65.8.1",
    "django-cors-headers>=4.7.0",
    "django-debug-toolbar>=5.2.0",
    "django-extensions>=4.1",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "drf-yasg>=1.21.10",
    "jwt>=1.3.1",
    "pillow>=11.2.1",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "rich>=14.0.0",
]
